Separated the second homework in different package
