+++ src/main/java/HW2/model/BaseDTO.java
@@ -1,0 +1,8 @@
+package HW2.model;
+
+public interface BaseDTO {
+
+    double getNumber();
+
+    void setNumber(double number);
+}
