+++ src/main/java/HW2/thread/BaseThread.java
@@ -1,17 +1,17 @@
-package thread;
+package HW2.thread;
 
-import model.BaseDTO;
-import model.FirstDTO;
-import model.SecondDTO;
+import HW2.model.BaseDTO;
+import HW2.model.FirstDTO;
+import HW2.model.SecondDTO;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import utils.CustomTimer;
+import HW2.utils.CustomTimer;
 
 import java.util.concurrent.locks.ReentrantLock;
 
 import static java.lang.String.format;
-import static utils.CustomRandom.getDoubleRand;
-import static utils.CustomRandom.getIntRand;
+import static HW2.utils.CustomRandom.getDoubleRand;
+import static HW2.utils.CustomRandom.getIntRand;
 
 public abstract class BaseThread extends CustomTimer implements Runnable {
