--- src/main/java/model/BaseDTO.java
@@ -1,8 +1,0 @@
-package model;
-
-public interface BaseDTO {
-
-    public double getNumber();
-
-    public void setNumber(double number);
-}
